# Implement host rotation and Enqueue/Dequeue access regulation via atomic booleans