--- a/internal/pkg/queue/access.go
+++ b/internal/pkg/queue/access.go
@@ -1,0 +1,9 @@
+package queue
+
+func (q *PersistentGroupedQueue) CanEnqueue() bool {
+	return !q.closed.Get()
+}
+
+func (q *PersistentGroupedQueue) CanDequeue() bool {
+	return !q.closed.Get() && !q.finishing.Get()
+}