@equals215 2024-07-26
@CorentinB can you also check if those changes won't interfere with the way we enqueue items from HQ?
I have to go for tonight